update annex environment section
Updating the ANNEX document to have a sub-section dedicated for listing
environment options that releng-tool accepts (over script context-
provided options). This change adds the new section and adds the
existing set of applicable environment variables. This also removes
a subset of variables listed in the context group which was in err.

Signed-off-by: James Knight <james.d.knight@live.com>