use cascadia code as default monospace font to improve block elements
With a non-one line-height configured, the default theme fonts used do
not handle the rendering of block elements well. Through some visual
inspection, the `Cascadia Code` font looks to be a font that renders
block elements well with the larger line height. Switching to use it as
the default for pre-elements to improve our file-structure diagrams.

Signed-off-by: James Knight <git@jdknight.me>